requirements.txt
discord
python-dotenv
pytz